apply mobx and add SignUpPhoneStore and delete PhoneCodeAuthStore
