+++ app/src/constant/TimeOut.js
@@ -1,0 +1,4 @@
+export const TIME_OUT = {
+  START_TIME: '180',
+  END_TIME: '0',
+};
