+++ app/src/repository/FirebaseRepository.js
@@ -7,7 +7,6 @@ const TARGET_URL = `${SERVER_URL}/users`;
 export default class FirebaseRepository {
   sendSignUpPhoneCode = async (phoneNumber) => {
     const confirmation = await auth().signInWithPhoneNumber(phoneNumber);
-    auth().verifyPhoneNumber
     return confirmation;
   };
 }
